+++ pkgs/test_api/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Deprecate `throwsNullThrownError`, use `throwsA(isA<TypeError>())` instead.
   The implementation has been changed to ease migrations.
+* Deprecate `throwsCyclicInitializationError` and replace the implementation
+  with `Throws(TypeMatcher<Error>())`. The specific exception no longer exists
+  and there is no guarantee about what type of error will be thrown.
 
 ## 0.4.16